// maia-showcase/maia-showcase-ui/src/generated/typescript/main/app/gen-components/org/maiaframework/showcase/all-field-types/all-field-types-gen-routes.ts
// This source was generated by the Maia Framework code generator
// Renderer class: class org.maiaframework.gen.renderers.ui.GenRoutesRenderer

import {Routes} from '@angular/router';


export const allFieldTypesGenRoutes: Routes = [
    {
        path: 'all-field-types-blotter',
        loadComponent: () =>
            import('./all-field-types-blotter-page').then(m => m.AllFieldTypesBlotterPage),
    },
    {
        path: 'all-field-types/view/:id',
        loadComponent: () =>
            import('./all-field-types-entity-detail-view-page').then(m => m.AllFieldTypesEntityDetailViewPage),
    },
    {
        path: 'all-field-types/create',
        loadComponent: () =>
            import('./all-field-types-entity-create-page').then(m => m.AllFieldTypesEntityCreatePage),
    },
    {
        path: 'all-field-types/edit/:id',
        loadComponent: () =>
            import('./all-field-types-entity-edit-page').then(m => m.AllFieldTypesEntityEditPage),
    },
];