// maia-showcase/maia-showcase-ui/src/generated/typescript/main/app/gen-components/org/maiaframework/showcase/join/bravo-gen-routes.ts
// This source was generated by the Maia Framework code generator
// Renderer class: class org.maiaframework.gen.renderers.ui.GenRoutesRenderer

import {Routes} from '@angular/router';


export const bravoGenRoutes: Routes = [
    {
        path: 'bravo-blotter',
        loadComponent: () =>
            import('./bravo-blotter-page').then(m => m.BravoBlotterPage),
    },
    {
        path: 'bravo/view/:id',
        loadComponent: () =>
            import('./bravo-entity-detail-view-page').then(m => m.BravoEntityDetailViewPage),
    },
    {
        path: 'bravo/create',
        loadComponent: () =>
            import('./bravo-entity-create-page').then(m => m.BravoEntityCreatePage),
    },
    {
        path: 'bravo/edit/:id',
        loadComponent: () =>
            import('./bravo-entity-edit-page').then(m => m.BravoEntityEditPage),
    },
];