// maia-showcase/maia-showcase-ui/src/generated/typescript/main/app/gen-components/org/maiaframework/showcase/composite-pk/composite-primary-key-gen-routes.ts
// This source was generated by the Maia Framework code generator
// Renderer class: class org.maiaframework.gen.renderers.ui.GenRoutesRenderer

import {Routes} from '@angular/router';


export const compositePrimaryKeyGenRoutes: Routes = [
    {
        path: 'composite-primary-key-blotter',
        loadComponent: () =>
            import('./composite-primary-key-blotter-page').then(m => m.CompositePrimaryKeyBlotterPage),
    },
    {
        path: 'composite-primary-key/view/:id',
        loadComponent: () =>
            import('./composite-primary-key-entity-detail-view-page').then(m => m.CompositePrimaryKeyEntityDetailViewPage),
    },
    {
        path: 'composite-primary-key/create',
        loadComponent: () =>
            import('./composite-primary-key-entity-create-page').then(m => m.CompositePrimaryKeyEntityCreatePage),
    },
    {
        path: 'composite-primary-key/edit/:id',
        loadComponent: () =>
            import('./composite-primary-key-entity-edit-page').then(m => m.CompositePrimaryKeyEntityEditPage),
    },
];