// maia-showcase/maia-showcase-ui/src/generated/typescript/main/app/gen-components/org/maiaframework/showcase/many-to-many/left-many-gen-routes.ts
// This source was generated by the Maia Framework code generator
// Renderer class: class org.maiaframework.gen.renderers.ui.GenRoutesRenderer

import {Routes} from '@angular/router';


export const leftManyGenRoutes: Routes = [
    {
        path: 'left-many-blotter',
        loadComponent: () =>
            import('./left-many-blotter-page').then(m => m.LeftManyBlotterPage),
    },
    {
        path: 'left-many/view/:id',
        loadComponent: () =>
            import('./left-many-entity-detail-view-page').then(m => m.LeftManyEntityDetailViewPage),
    },
    {
        path: 'left-many/create',
        loadComponent: () =>
            import('./left-many-entity-create-page').then(m => m.LeftManyEntityCreatePage),
    },
    {
        path: 'left-many/edit/:id',
        loadComponent: () =>
            import('./left-many-entity-edit-page').then(m => m.LeftManyEntityEditPage),
    },
];