// maia-showcase/maia-showcase-ui/src/generated/typescript/main/app/gen-components/org/maiaframework/showcase/simple/simple-gen-routes.ts
// This source was generated by the Maia Framework code generator
// Renderer class: class org.maiaframework.gen.renderers.ui.GenRoutesRenderer

import {Routes} from '@angular/router';


export const simpleGenRoutes: Routes = [
    {
        path: 'simple-blotter',
        loadComponent: () =>
            import('./simple-blotter-page').then(m => m.SimpleBlotterPage),
    },
    {
        path: 'simple/view/:id',
        loadComponent: () =>
            import('./simple-entity-detail-view-page').then(m => m.SimpleEntityDetailViewPage),
    },
    {
        path: 'simple/create',
        loadComponent: () =>
            import('./simple-entity-create-page').then(m => m.SimpleEntityCreatePage),
    },
    {
        path: 'simple/edit/:id',
        loadComponent: () =>
            import('./simple-entity-edit-page').then(m => m.SimpleEntityEditPage),
    },
];