// maia-showcase/maia-showcase-ui/src/generated/typescript/main/app/gen-components/org/maiaframework/showcase/versioned/some-versioned-gen-routes.ts
// This source was generated by the Maia Framework code generator
// Renderer class: class org.maiaframework.gen.renderers.ui.GenRoutesRenderer

import {Routes} from '@angular/router';


export const someVersionedGenRoutes: Routes = [
    {
        path: 'some-versioned-blotter',
        loadComponent: () =>
            import('./some-versioned-blotter-page').then(m => m.SomeVersionedBlotterPage),
    },
    {
        path: 'some-versioned/view/:id',
        loadComponent: () =>
            import('./some-versioned-entity-detail-view-page').then(m => m.SomeVersionedEntityDetailViewPage),
    },
    {
        path: 'some-versioned/create',
        loadComponent: () =>
            import('./some-versioned-entity-create-page').then(m => m.SomeVersionedEntityCreatePage),
    },
    {
        path: 'some-versioned/edit/:id',
        loadComponent: () =>
            import('./some-versioned-entity-edit-page').then(m => m.SomeVersionedEntityEditPage),
    },
];