// maia-showcase/maia-showcase-ui/src/generated/typescript/main/app/gen-components/org/maiaframework/showcase/user/user-gen-routes.ts
// This source was generated by the Maia Framework code generator
// Renderer class: class org.maiaframework.gen.renderers.ui.GenRoutesRenderer

import {Routes} from '@angular/router';


export const userGenRoutes: Routes = [
    {
        path: 'user-blotter',
        loadComponent: () =>
            import('./user-blotter-page').then(m => m.UserBlotterPage),
    },
    {
        path: 'user/view/:id',
        loadComponent: () =>
            import('./user-entity-detail-view-page').then(m => m.UserEntityDetailViewPage),
    },
    {
        path: 'user/create',
        loadComponent: () =>
            import('./user-entity-create-page').then(m => m.UserEntityCreatePage),
    },
    {
        path: 'user/edit/:id',
        loadComponent: () =>
            import('./user-entity-edit-page').then(m => m.UserEntityEditPage),
    },
];